Add support to work around ExpandExe problems
So ExaGear Strategies does not support GetThreadContext properly. Changed so it can detect this and then does not try and expand the exe.
Also, added a new option that can be added "DontExpandExe = true" which will force the exe not to be expanded.
Also made it so it reads the .ini file last line even without carriage return